fix: corregir issues del code review
- isVerifyingBalance: timeout de 30s con Handler en lugar de depender del
  primer SMS de saldo. Evita que mensajes no relacionados interfieran y
  garantiza que el flag siempre se resetea aunque no llegue respuesta
- isVerifyingBalance: seteado ANTES del check de permisos para que persista
  a través del flujo de permisos en onRequestPermissionsResult
- isVerifyingBalance: se resetea también si el mensaje de 226 llega pero
  no se puede parsear el saldo (antes quedaba colgado forever)
- showManualConfirmationDialog: recalcula pending al momento del click en
  lugar de capturarlo al abrir el diálogo (evita confirmar de más si llegan
  SMS de confirmación mientras el diálogo está abierto)
- MonthlyTotal.total: Int → Long para no truncar acumulados grandes de SQLite
- Escapar $ en string templates Kotlin (manuallyConfirmPending Toast e historial)
- updateReminders: agregar requestCode 2 al cancel loop para migración desde
  el esquema anterior (evita alarms huérfanos en dispositivos existentes)
- Touch targets: historyButton y manualConfirmButton usan minHeight=48dp
- onDestroy: cancelar el timeout handler para evitar fugas

+++ app/src/main/java/uy/roar/donadorautomatico/data/DonationDao.kt
@@ -8,7 +8,7 @@ import androidx.room.Query
 
 data class MonthlyTotal(
     @ColumnInfo(name = "month") val month: String,
-    @ColumnInfo(name = "total") val total: Int
+    @ColumnInfo(name = "total") val total: Long
 )
 
 @Dao
